refactor: LineService 의 findResponseById() 에도 역의 순서대로 노출되도록 적용
