+++ src/test/java/nextstep/subway/section/SectionAcceptanceTestActions.java
@@ -12,7 +12,7 @@ import nextstep.subway.domain.Station;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.MediaType;
 
-public abstract class SectionTestFixtures {
+public abstract class SectionAcceptanceTestActions {
 
     private static final String PATH_LINE = "/lines";
     private static final String PATH_LINE_PATH_LINE_ID = PATH_LINE + "/{lineId}";
@@ -68,27 +68,8 @@ public abstract class SectionTestFixtures {
 
     public static void 등록_완료(String pathVariable, Station... stations) {
         JsonPath 목록조회결과 = 목록조회(PATH_LINE_PATH_LINE_ID, pathVariable);
-        assertThat(목록조회결과.getList("stations.name", String.class))
-                .contains(stations[0].getName(), stations[1].getName(), stations[2].getName());
-    }
-
-    public static void 기존_구간_뒤에_하행_종점으로_등록한_중앙역_구간이_함께_조회됨(String pathVariable, String... containValues) {
-        JsonPath 목록조회결과 = 목록조회(PATH_LINE_PATH_LINE_ID, pathVariable);
-//        assertThat(목록조회결과.getList("distances", String.class)).containsExactly(containValues);
-        assertThat(목록조회결과.getList("stations.name", String.class)).contains("경기 광주역", "모란역", "중앙역");
-    }
-
-
-    public static void 기존_구간_앞에_상행_종점으로_등록한_모란역_구간이_함께_조회됨(String pathVariable, String... containValues) {
-        JsonPath 목록조회결과 = 목록조회(PATH_LINE_PATH_LINE_ID, pathVariable);
-        assertThat(목록조회결과.getList("distances", String.class)).containsExactly(containValues);
-        assertThat(목록조회결과.getList("stations.name", String.class)).contains("모란역", "경기 광주역", "중앙역");
-    }
-
-    public static void 새로운_길이를_뺀_나머지를_새롭게_추가된_역과의_길이로_설정(String pathVariable, String... containValues) {
-        JsonPath 목록조회결과 = 목록조회(PATH_LINE_PATH_LINE_ID, pathVariable);
-//        assertThat(목록조회결과.getList("distances", String.class)).containsExactly(containValues);
-        assertThat(목록조회결과.getList("stations.name", String.class)).contains("경기 광주역", "모란역", "중앙역");
+        assertThat(목록조회결과.getList("stationNames", String.class))
+                .containsExactly(stations[0].getName(), stations[1].getName(), stations[2].getName());
     }
 
     public static ExtractableResponse<Response> 새로운_역_하행_종점으로_등록(String upStationId, String downStationId,
